megascrapper/files/requirements.txt
requests
datetime
influxdb-client
google-cloud-storage
google-cloud-secret-manager
pyyaml
